Improve Cutoffs, KB Weighting and data quality on llm websearch (#40)
* Improved classifications by adding gics information

* Improved Assessment score calculation and KB exctractions

* enhanced assessment score calculation

* Enhanced data collectioon by automated refinement runs

* Update documentation

+++ services/app/backend/src/main/java/my/portfoliomanager/app/model/LayerTargetRiskThresholds.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class LayerTargetRiskThresholds {
-	private final Integer lowMax;
-	private final Integer highMin;
+	private final Double lowMax;
+	private final Double highMin;
 }
